multicast: Adding new folder for test tools
The `test_tools/' directory is a place for storing
various non-python tools that will later be used for
testing, such as programs in C, shell scripts etc.

This commit also adds first test tools - a set of
utilities, test setups and cases for validating
implementation of multicast/igmp in kernel.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>