Add skb_clone option to PktgenTx test
The pktgen's skb_clone > 0 is not valid for bonds/vlans in RHEL. This is
due to fix of bug https://bugzilla.redhat.com/show_bug.cgi?id=678794.

Following patch adds skb_clone option that can be set to 0 for such cases.
The default value remains 100000. Tested by myself on RHEL6.