NetTestParse: Adding XML Preprocessor
Code that handles parsing aliases was refactored and moved from
NetTestParse into a new class in Common module - XmlPreprocessor.

In the process of this transition it was extended with a new type
of template - functions.

Template functions can be used the same way as aliases - inside of
curly braces. For instance:

    {ip(1,2,3)}
    {function(argument1, $alias_argument)}

There are three template functions available at the moment. All three
can be used for retrieving information from machine configs:

    hwaddr(machine_id, if_id) - get hw address of an interface
    devname(machine_id, if_id) - get interface name
    ip(machine_id, if_id, addr_id) - get IP address (if there's a mask
                                     present in the configuration it
                                     will be stripped). The third
                                     argument is optional, defaults to
                                     zero if omitted.

Aliases can be passed to functions as parameters, but the preprocessor
does not support nesting (e.g. function cannot have another function
passed as an argument).

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>