Do not stop command sequence execution if command failed
In case of a failed sequence command (even test that did not pass) the
whole rest of the command sequence was skipped. This could result in
hanging background process that would not be optionally killed at the
end of command sequence using type="kill" command. The fix is to let
the whole command sequence continue and save any failed command for
the overall command sequence result.