NetConfigParse: Extending options parsing
This commit makes NetConfigParser accept <netdevice> options in a
slightly different format. Previous format of options accepted values
only as an 'value' attribute to the <option> tag. With this modification
it is possible to define options also with values as a text elements of
the option tag. There are now two possibilities:

    <!-- The usual way -->
    <option name="opt" value="5" />

    <!-- Longer values can be specified like this -->
    <option name="teamd_config">
        {
            "device":   "team1",
            "runner":   {"name": "roundrobin"},
            "ports":    {"eth3": {}, "eth4": {}}
        }
    </option>

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>