XmlPreprocessor: Modify template detection
The previous solution considered curly braces to be reserved
characters, exclusively for marking templates. But according
to recent development, it will be necessary to deal with
those characters also in non-template context.

This commit enhances template detection code so it will ignore
any occurencies of curly braces unless they form a valid template.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>