Tests: Adding new test - PacketAssert
This commit introduces a new test - packet assert, that can be used to
test if a specific amount of packets with certain properties were
transfered through some interface.

Packet properties can be verified through the pcap "filter" option or
through greping tcpdump output by using "grep-for" option (grep-for can
be used multiple times.

Here's a quick examle:
    <command type="test" value="PacketAssert" bg_id="1">
        <options>
            <option name="interface" value="br0" />
            <option name="filter" value="dst port 1234" />
            <option name="grep_for" value="IP" />
            <option name="min" value="1" />
            <option name="max" value="5" />
        </options>
    </command>
    <command type="exec" value="sleep 3" />
    <command type="intr" value="1" />

This assert evaluates true if one to five IP packets with destination
port of 1234 are transfered through interface br0 within the time period
of sleep (i.e. 3 seconds).

PacketAssert must always be run in background. It is evaluated at the
time of interruption with "intr". It will not evaluate correctly in
case the command is killed (with "kill")!

One example recipe to demonstrate it's capabilities was added as well.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>