nettestctl: New CLI option --packet_capture (-p)
This commit introduces new CLI `-p' option to nettestctl.py.
If enabled (it is off by default), network communication from all
test interfaces (excluding controller ports) are logged to hard-drive.

These logs are transfered to controller machine when the recipe
exection is over. In case of an error at the slave side, the dump
files are still transfered, but they will not be transfered when
the controller is killed during its execution.

Log files are in binary *.pcap format and can be found in the `Logs'
directory with the slave log files. For instance:

    Logs/2012-06-20_18:56:42/mcast/10.34.1.236/1.pcap

    The path contains:
        1. date of execution - 2012-06-20_18:56:42
        2. recipe name - mcast
        3. slave hostname - 10.34.1.236
        4. capture file name - 1.pcap

The naming convention of the capture file is <id>.pcap, where <id>
is a netdevice id from machine's netconfig.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>