Use modinfo instead of deprecated modprobe
Since modprobe -l is deprecated and modinfo should be used instead
I made the update to type_check method.

I have tested it on RHEL5 and RHEL6. On RHEL5 I've hit another issue
on macvlan check. The module is not available on RHEL5 and the slave
instance failed due to die_on_err=True (default). I changed this to
False since we might have cases when the module for a network device
is not available.