nettestctl.py: all_dump and eval actions removed
'all_dump' now makes no sense. Configuration of slave machines now
occurs at recipe parsing time (because it's needed to be done for
further processing). That means, that you cannot safely parse the
recipe config without connecting to the slaves.

'eval' option is tied to the recipe_eval functionality that was
removed and replaced by aliases and templates.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>