NetTestParse: New parser implementation
This patch replaces the original parser with another one that operates
on a different principle. It parses tags in order of their occurence
in the source XML file.

Some other features were added to the parser as well. For instance,
support of advanced error reporting that now includes location of
where the error occured (the exact file name and line number).

The parser is able to export some predefined events, that can be
used to signal the parent code. Parent can register handlers for
these events.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>