NetTestController: Fix behavior of -c parameter
I "fixed" what I thought was a bad behavior of the cleanup option in
commit 791fd48e6a95a10dfef60ac442f2a34f225e672a [1]. As it turns out,
my understanding of the purpose of -c option was wrong and the current
behavior is in fact a regression.

This commit reverts the changes and takes the semantics of '-c' option
back to the original.

If you pass -c option to `nettestctl.py', the machines will be cleaned
up *before* the recipe is executed. The option now has no effect on the
cleanup, that takes place after the recipe was run. This is done always
automatically.

[1] NetTestController: Adding missing cleanup
    https://fedorahosted.org/pipermail/lnst-developers/2012-June/000149.html

Reported-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Radek Pazdera <rpazdera@redhat.com>