NetTestParse: Changing 'netdevice' to 'interface'
This patch changes the name of 'netdevice' tag within 'netconfig'
to 'interface' to make it more distinguishable from 'netdevice' tag
that is also present within 'netmachineconfig', but with a little
different semantics. I posted more in-depth description of this
to the lnst-developers list [1].

This change should help users to understand the XML recipes better.

I also tried shorter 'iface' instead of 'interface' (as we discussed
at the meeting), but in my opinion it doesn't fit there right. None
of the other tag names is abbreviated and 'interface' is 9 chars long,
which is exactly the same as the previous 'netdevice'.

Here's an example:

    <netconfig>
        <interface id="1" phys_id="1" type="eth"/>
        <interface id="2" phys_id="2" type="eth"/>
        <interface id="3" type="bond">
            <options>
                <option name="mode" value="0" />
                <option name="miimon" value="100" />
            </options>
            <slaves>
                <slave id="1"/>
                <slave id="2"/>
            </slaves>
            <addresses>
                <address value="192.168.122.3/24" />
            </addresses>
        </interface>
    </netconfig>

  [1] https://fedorahosted.org/pipermail/lnst-developers/2012-July/000219.html

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>