NetConfig: Small code cleanup
This commit merges type_init() function with add_interface_config() and
type_cleanup() with remove_interface_config().

These functions had too dangerous preconditions, which was not very
obvious. The preconditions disappear after the merge.

Also an info entry will be written to the log when a device class is
initialized or cleaned up.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>