ExecCmd: Adding some info to ExecCmdFail exception
This commit extends ExecCmdFail exception with some information of
what went wrong, so the parent code that can catch this error can
react appropriatelly.

ExecCmdFail now contains return value of the command, the command
itself (for reporting purposes - which command failed) and conents
of standard error output, that usualy contains a clue or some error
message.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>