NetTest: Adding 'network' attribute to netdevice
This commit introduces networks into LNST netdevices. Up to this point
we were unaware of how the devices are connected to each other. The
network attribute adds this information directly into the recipe.

So if you later receive a recipe, you can configure your underlying
physical network accordingly. This will be also useful with the
upcomming virtual networks and interfaces, because LNST will be able
to set up the virtual network automatically.

See the following example:

  Machine 1:
    <netdevice hwaddr="52:54:00:11:D5:A8" network="channel-1"
               phys_id="1" type="eth"/>
    <netdevice hwaddr="52:54:00:2E:CD:A8" network="channel-2"
               phys_id="1" type="eth"/>

  Machine 2:
    <netdevice hwaddr="52:54:00:2E:D5:A8" network="channel-1"
               phys_id="1" type="eth"/>
    <netdevice hwaddr="52:54:00:2E:D5:A8" network="channel-2"
               phys_id="1" type="eth"/>

Machine 1 is connected to machine 2 with two networks, channel-1 and
channel-2.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>