NetTestParse: Adding support for libvirt-integration
This commit adds necessary changes to the parser in LNST in order
to support the libvirt device creation.

Netdevice tags are now grouped under a <netdevices> parent tag
within <netmachineconfig>.

On top of that a special mode for defining netdevices (called create)
was added. Devices defined in this mode are not present in the machine
and will be created on test execution.

This mode is available only on virtual test machines created using
libvirt. Here's an exmaple config of such a machine:

<machine id="1">
    <netmachineconfig>
        <info hostname="192.168.122.11" libvirt_domain="RHEL6"
              rootpass="redhat"/>
        <netdevices>
            <libvirt_create>
                <netdevice network="tnet" phys_id="1" type="eth"/>
                <netdevice network="tnet" phys_id="2" type="eth"/>
            </libvirt_create>
        </netdevices>
    </netmachineconfig>
    <netconfig>
        <interface id="1" phys_id="1" type="eth">
            <addresses>
                <address value="192.168.100.226/24"/>
            </addresses>
        </interface>
        <interface id="2" phys_id="2" type="eth">
            <addresses>
                <address value="192.168.100.240/24"/>
            </addresses>
        </interface>
    </netconfig>
</machine>

The dynamic devices are enclosed within the <libvirt_create> tag. Their
hwaddr doesn't need to be specified (it will be generated if missing).

Additionally, 'libvirt_bridge' attribute can be added to these devices
to override the default LNST internal bridge, that these devices are
connected to.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>