system_config: Multiple values for single option support
This commit fixes bug (trac #3) in system config code, namely the
NetTestCommandSystemConfig class.

In cases when value was assigned to a same option multiple times
within a single system_config command, only the first value was
used. The remaining ones were discarded.

Now the option is set one by one to all the values as they are
specified in the command.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>