Ticket #4: machine_id from int to string
Machine ids in tags <machine id=''> and <command machine_id=''> are no
longer converted to integers and are used as strings.
Introducing these changes requires a number of changes in formatted
strings where integers were required.
Machine id is also used as the first parameter in template functions so
these had to be changed as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>