Ticket #4: interface_id from int to string
Interface ids in tags <interface id=''> and <slave id=''> are
no longer converted to integers and are used as strings.
Introducing these changes requires a number of changes in formatted
strings where integers were required.
Interface id is also used as the second parameter in template functions
so these had to be changed as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>