NetTestController: Write traceback to debug on error
When the recipe execution fails with an exception in NetTestController,
the exception is always caught and rethrown after propper cleanup has
been done.

This unfortunatelly leads to loosing the original exception traceback
which is very helpful for debugging.

The original traceback is now logged into debug when exception is
caught and then rethrown.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>