NetUtils: add verify functions for ip and mac addresses
This commit adds functions verify_ip_address and verify_mac_address that
return a boolean value depending on the validity of given ip or mac
addresses.

These will be used in the new Config parsing module.
I also added their usage to the address generation code in the same
module.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>