Common: Adding config module
This patch adds a module for parsing config files.
The class Config uses the module ConfigParser for parsing .ini files.
Parsed data is then read and checked for invalid sections, options or
option values.
The main method for this class is load_config() where we check for
invalid sections. Next in methods section*() we check for invalid
options. Finally values are checked in methods option*().

Correctly parsed options are stored in a dictionary
options[section][option].

It is important to add that parts of this class are not fully complete
and it is highly possible more sections/options will be added during the
process of integrating this class with the rest of the project.
For the same reason some option methods are there only as place holders
until they can be fully implemented.

Default values should be stored in a separate config file. For now I am
using default.conf that is located in the main project directory. I
don't know if this is the best way so I am not including this file in
this commit.

Currently the accepted configuration files are of this format, you can
use this as the default.conf file:

[environment]
mac_pool_range = 52:54:01:00:00:01 52:54:01:FF:FF:FF
rpcport = 9001
[log]
local_ip = 127.0.0.1
port = 9000
log_path = ~/.lnst/logs

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>