Config: cleanup of class Config
In this patch I removed the support for option 'local_ip' ip section
'log' because this value is only used in one place as part of an output
string.

I also added proper initialization of the data structure that stores
loaded configuration values. The default values used here are values
that were previously used as global variables. Functionality therefore
shouldn't change.

The parser is now a local object of the function load_config, previously
it would cause problems by using old values when nonexistant files were
to be parsed.

I also added exception handling to functions get_section and get_option.

Finally I modified the way paths are handled- relative paths should be
relative to the location of the config file they are specified in. And
paths using ~ should be properly expanded to the current users home.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>