add lnst.conf.example configuration file
This file is being added as an example of how to create custom
configuration files. For illustration it contains the default values
currently used by lnst.
All that needs to be done is changing the values and place the resulting
file in a proper location. Currently only file ~/.lnst/lnst.conf will be
used.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>