MachinePool: get pool dirs from config
This patch adds support for option 'machine_pool_dirs' in section
'environment' of config files.
This option is a list of directories separated by whitespace.
Whitespaces in names need to be escaped with '\'.
Behaviour for this option is also a bit different, config files with
higher priority don't overwrite the previous value but instead are added
to the list.

This list is later passed to class MachinePool. Currently this class
doesn't check for the existence of these directories. Nonexistent
directories will cause a crash of the application so be sure to create
them before trying to use them.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>