Logs: warn user if slave can't connect to logging server
Currently logging doesn't work correctly if the connection between a
slave and the controller isn't properly set up. This can be for example
caused by a firewall. In these cases it can be confusing to users
because there is no indication of it other than missing logs.

This patch changes that. It adds a new function test_tcp_connection() to
the NetUtils module. Slaves use it to check if they are able to connect to
the logging server and inform the controller of the results. The
controller can then log an error message in case of a failure.
I tried making the error message more visible so that it's harder to
overlook.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>