XmlProccessing: no password bug fix
Machine passwords are optional, however not specifying one causes the
controller crash with an undescriptive error message. This is caused by
the fact that in these cases None value is used instead of the password.

This commit changes that by using an empty string as the password when
it isn't specified in the recipe.

I also changed the error message for failed password authentification,
so that it is more clear that the password failed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>