logging: add new file LoggingHandler
This file defines a new class ServerHandler. This class is a logging
handler and therefore is a decendant of the class logging.Handler.

This new class contains a TCP server socket listening on a specified
port. Until an incoming connection is recieved emitted messages are
stored in a buffer. After recieving a connection the stored messages are
all sent to the client and the buffer is flushed.

To avoid using threads or a new process for this server, the server
socket is created as nonblocking and we check for incoming connections
only when a new log message is emitted.

This handler was created because there is no suitable alternative for
this functionality in the python standard library. However this handler
is inspired by the SocketHandler and uses some of it's code.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>