LoggingServer: invert the direction of connections
This commit changes the LoggingServer so that it doesn't listen for
incoming connections and instead initiates them. This solves the
problem of slave logs not arriving when the controller is using a
firewall or is located behind a NAT.

To connect a new slave its ip address and the port number are sent to
the server process through a pipe. The server then attempts to create a
new connection. This expects that there is a listening server socket on
the other side, which for us means that the slave is using the
ServerHandler.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>