Logs: slaves use Server handler instead of SocketHandler
This commit makes slaves use the newly implemented ServerHandler instead
of a combination of a MemoryHandler and a SocketHandler. This is another
one of the changes required for the reversal of the communication
between controller and slave logging.

I also removed the function append_network_hadler since it won't be used
from now

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>