controller: changes required by the new LoggingServer
NetTestController now needs to have access to the logging server so
nettestctl now passes it as an argument to the constructor.

The function _init_slave_logging now doesn't communicate with the slave
at all. Instead it sends the slaves address to the logging server which
then tries to connect to it.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>