MachinePool: Adding provisioning support
The stub implementation for MachinePool class was extended with
a full provisioning support. It can be now used to supply machine
setups for requests from the NetTestController.

A new class -- SetupMapper was added to aid MachinePool in the matter
of matching requested setups against the pool.

SetupMapper is able to find a mapping between a requested topology
and available pool resources. The resulting setup can be then used
to run tests.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>