ExecCmd: Adding new parameter to exec_cmd
This new boolean parameter allows the caller to specify whether
to include the stderr output of the executed command in the exception
that is thrown.

Some commands have very nice and descriptive one-line error reports
that LNST can pass along in the exception and print to user. On the
other hand, some different commands print huge error reportings or
memory dumps that definitely should not be printed to the user.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>