XmlRpc: Dispatch error messages not tracebacks
When a remote RPC method raises an exception it's traceback is
marshalled back to the controller without the error message that
the exception carried. The traceback is suitable for debuging,
but it is certainly not suited for users.

This patch makes the dispatcher log the traceback to debug first
and then send only the error message. The traceback is transfered
to controller through the logging facility.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>