LoggingHandler: add a new handler
This patch adds a new handler that is used for storing logged messages
so that they can be retrieved later.

It also replases the ServerHandler that was used by the slave machines
to send logs to the controller.

For now the Server handler can stay as I don't know if it might come in
handy later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>