NetTestController: retrieve slave logs through rpc
The RPC server on the slaves now supports a new method get_new_logs that
is used for retrieving stored logs on the slave machine.

The controller now uses this method to retrieve the logs. Retrieved logs
are then passed to the method _add_client_logs that modifies the source
address and logges them on the controller.

The method _init_slave_logging had to be changed so that the directory
hierarchy for logs wouldn't be changed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>