Logs: small path change
The path where logs are located is now exactly the same as what you
specify in the configuration file. Previously 'Logs' was added to the
end of this path. You now need to do this manually.
This change allows for the removal of the log_root argument of the class
Logs. Instead log_folder should be used now.

The default path where you can find logs is <lnst_dir>/Logs .

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>