Config: support for different config schemes
The Config class now accepts an argument which specifies which
configuration scheme should be used. Currently I added a scheme
'controller', which was used until now as the configuration for the
controller, and 'slave' which will be used by the slave. The details of
the slave scheme will be added later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>