Config: change in hierarchy of config files
This patch changes the hierarchy by which we load configuration files to
the following:

1. We check for the existence of a configuration file located in the
same directory as the launched binary. This is for cases when the
application is launched from the git version. This configuration file
MUST NOT exist otherwise.
This configuration file has the HIGHEST priority and if it exists other
configuration files will not be loaded. This is to ensure that we can
run the git version and have the packaged version installed at the same
time without unnecessary conflicts.

2. We load the configuration file in /etc/ this should be the
system-wide configuration. We need this file mostly because of the slave
application which will be running as a daemon in the future. For
convention purposes and to not confuse the users this file will exist
even on the controller machines.
This file should be heavily documented and explain the available
options.

3. On the controller we load the configuration file located in the
~/.lnst directory. This can't be done on the slave machines as the slave
application will be running as a daemon and it would make sense to put
its configuration into the home directory of any user.

This patch also changes the names of the configuration files to:
lnst-ctl.conf
lnst-slave.conf
to distinguish the different schemes used in these files.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>