Config: add new options, update lnst-ctl.conf
This patch adds 2 new options:
'test_module_dirs' which specifies a list of dirs where to look for
    Test*.py files
'test_tool_dirs' which specifies a list of dirs where to look for
    test_tool packages to be used by slaves

We agreed to use += for some options that work with lists. After this
patch, options that specify a list of values can be used with both '='
and '+='.

To specify a new list replacing the list loaded from previous
configuration files use '=', e.g.
'machine_pool_dirs = ~/.lnst/pool1/ ~/.lnst/pool2/'

To append to the values loaded from previous configuration files use
'+=', e.g.
'machine_pool_dirs += ~/.lnst/pool/'

This patch also updates the lnst-ctl.conf file to the current format.
The default values for the new options will be added later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>