Config: add slave option cache_dir
The slave configuration now accepts section '[cache]' and an option
'cache_dir' under it. This option controls where the slave will look for
and store test modules and test tools that we will be transmitting
through our rpc connection.

This patch also changes the lnst-slave.conf to reflect this change. The
default value used for this option is './cache' which will be used by
the git version of lnst.

There is also a minor change in the name of the method optionLogPath to
just optionPath as it isn't related to logs directly.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>