NetTestCommand: Adding caching support
This patch changes NetTestCommand so it can access tests stored
in the cache rather than directly from the tree.

The resource table with paths to all available modules and tools
is passed to the command.

Also new feature of command type exec was added. It can now be
issued with 'from' attribute that indicates what tools should
the command be executed from. For instance:

    <command machine_id="1" timeout="30" type="exec" from="multicast"
             value="./igmp_query" />

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>