NetTestController: Caching support implementation
This patch adds test caching support to the NetTestController class.

The new way of handling test resources implies, that controller can no
longer execute commands directly on itself, unless a configured slave
is runing on the same machine as well. Therefore the support for this
was removed in this patch as well.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>