xmlrpc: Modifying 'hello' and adding 'bye' method
This commit adds new 'bye' method -- an equivalent to 'hello'
method.

Hello is allways sent prior to any other method from controller to
slave. The bye method will be used to indicate that no method will
follow in this controller session.

The methods are used to initialize and cleanup some internal slave
structures and caches.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>