Config: reimplementation parsing hierarchy
This commit changes how we parse the config files. We still use the
ConfigParser for the direct parsing of the files, but the way we
transform its output is changed.

First of all the initialization chooses which scheme should be used.
This means that the options atribute is initialized in a particular way.
The main change is in the deepest point in the hierarchy of this
attribute. Previously we only stored the value of an option, but now
after this change the specific option is a dictionary consisting of
value, additive flag, method used for value handling, and the name used
in the configuration file.

This change allows us to replace the long switch statements for
something more compact. This means we no longer used separate methods
for different sections.

These changes however require all the option handling functions to
accept the same number of arguments. That means I had to add the
options cfg_path to some of these functions.

The reasoning behind this change is that if we are to be adding more and
more options the previous switch scheme would be getting more and more
unreadable. Another point is that the name of the option will be needed
to create the file itself at the first launch.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>