Config: add method dump_config
This patch adds a method that returns the currently loaded config values
in a .ini file formatted string. In the future it will be used to create
the config file in users home directory.

I decided to dump all config values and use the '=' operator. Using the
'+=' where it can be used would make more sense but it would be more
complicated to implement and I currently don't see any pretty way to do
it. But I might change this in the future.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>