nettestctl: create user config file at first launch
The controller detets its first launch by searching for a configuration
file in the users home directory. If it doesn't exist it creates it and
fills it with values that are currently loaded. Using this configuration
file is not yet fully functional as it requires the correct values to be
loaded from the system-wide configuration file /etc/lnst-ctl.conf which
will be added at a later date.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>