lnst-ctl: remove remote_exec option
This commit removes the lnst-ctl argument for remote execution of the
slave application. This option is handy for developers but would cause
problems when the application is installed as a package. The
functionality will be probably replaced by a wrapper script later.

The removal of this option makes a few changes in the NetTestController
file as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>