dist: Adding systemd service file
This commit adds two different things.

New dist/ directory for distribution specific things, such as spec files
for different distributions, packaging patches, init scripts. Really
anything that can be useful for distributing LNST as a package within
an operating system.

Also it adds a service file for systemd prepared for distribution
in Fedora to this directory.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>