Logs: multiple recipes bug fix
Somwhere during the changes to the logging I created a bug that makes it
that multiple recipes are logged into incorrect directories, getting
further into the folder hierarchy. A related problem that occured is
that the logs specific to a recipe would also contain the logs of all
recipes executed after it.

This patch fixes that by saving the paths and then restoring them when
moving on to the next recipe.

The bug was created by moving code from the LoggingServer to the
NetTestController. The LoggingServer had one static variable as the base
for the paths, however after moving it to the controller the variable is
no longer static and is changed by every slave we are connecting to.
This results in the next recipe being logged in the directory for the
last connected slave of the previous recipe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>