add command ctl_wait for the controller
Sometimes it is necessary for the controller to wait and do nothing
while the slaves do stuff in the background. Until now we were using the
exec type command. However recently we made a change that forbid the
controller to run commands.

This patch adds a new command type "ctl_wait" that accepts an integer
value indicating for how many seconds will the controller sleep. To
ensure the usage of this command only for the controller the only
accepted attributes are: type, value and desc. The attribute machine_id
is not supported since the command can only be used by the controller.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>