NetTestSlave: check if NetworkManager is running
This commit adds function check_process_running that checks if there is
a process with a specified name running on the system.

This function is then used in the slaves rpc method hello() to check if
NetworkManager is running. NetworkManager is know for causing problem s
during automatic network testing when he interferes with the testing
network interfaces. The slave now logs a warning message to point out
these potential problems.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>