NetTestController: check for running libvirtd
The controller now checks if the libvirt daemon is running and based on
this it decides if the MachinePool class should or shouldn't load
machine files that describe virtual machines.

There is also a minor change to the check_process_running function to
redirect the command output so that it isn't logged.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>