Fix logging issue with python2.6
This patch fixes logging issue that caused that the logs for a recipe
were appended with content of further recipes in the test run. This
issue was seen only with python2.6. It's type() built-in function did
not worked properly and python documentation recommends to use isinstance()
instead. This issue was not seen with python2.7.

Also, in case of passing recipe directory the typo was discovered and
fixed, root_logger was incorrectly written as logger.

Signed-off-by: Jan Tluka <jtluka@redhat.com>