Logging: fix for logs from bg processes
This commit fixes a bug in logging where there would be logs from
background commands. The bug is another one of bugs that were introduced
when we started retrieving logs through rpc. After this commit the logs
from background processes are retrieved through the result object that
we send through the created pipe.

This fix is not perfect because the logs from background commands will
be retrieved through their corresponding wait/intr/kill command. This
will result in logs from these commands not arriving at the moment of
their creation and if more than one background command was running then
the logs will not be in the "correct" order. However this fix is the
easiest and most "clean" of ways to fix this. We might consider better a
better way to do this when we redo the logging infrastructure.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>