nettestrecipe renaming to lnstrecipe
This patch renames the nettestrecipe tag to lnstrecipe as we discussed
in the meeting. I also renamed the corresponding method in
NetTestParse.py so that it reflects this change.

Please remember to change your custom recipes to reflect this change as
the old recipes will not work after application of this patch.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>