Logs: splitting slave logs into directories
Since slaves are now running as daemons they will be mostly run once for
multiple runs of the controller. This results in logs from individual
controller runs and recipes being logged in the same files on the slave
machines. This behaviour is confusing and unwanted.

This commit fixes that by making the slave logging change location every
time the rpc method hello() is called. For this I added another
classmethod to the class Logs- relocate_log_folder(). This function acts
in a similar way to the contructor of the class but only redirects the
paths. This pats makes the following logs structure:

.../<time>/{debug, info} - logs from the slave daemon unrelated to the
    recipe execution, currently only startup information, new folder is
    created every time the slave is launched
.../<time>_<recipename>/{debug,info} - logs generated by the result, new
    folder is created every time the hello() method is called

To avoid timestamp conflict of two identically named recipes being
launched in the same second, we wait for one second after the creation
of the log folder.

Finally I also removed the Logs class attribute file_handlers, it wasn't
used for anything and I probably forgot to remove it the last time I was
doing changes to the Logs class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>